requirements.txt
Django==3.1.5
djangorestframework==3.11.2

stripe==2.56.0

python-dotenv==0.15.0
psycopg2==2.8.6
django-cors-headers==3.7.0

black==20.8b1
flake8==3.8.3